Revert "Tipo de datos precio a double"
This reverts commit 07dcb2565a0253213cf1a8845b5d02e04d97bafa.

+++ proyecto-codeblocks-dev/include/Producto.h
@@ -11,7 +11,7 @@ class Producto
 public:
     //CONSTRUCTORES
     Producto();
-    Producto(int ID, int categoria, int idMarca, std::string modelo, std::string descripcion, double precioVenta, int stock, bool activo);
+    Producto(int ID, int categoria, int idMarca, std::string modelo, std::string descripcion, float precioVenta, int stock, bool activo);
 
     //SETTERS
     void setID(int ID);
@@ -19,7 +19,7 @@ public:
     void setMarca (int idMarca);
     void setModelo(std::string modelo);
     void setDescripcion(std::string descripcion);
-    void setPrecio (double precioVenta);
+    void setPrecio (float precioVenta);
     void setStock (int stock);
     void setActivo (bool activo);
 
@@ -29,7 +29,7 @@ public:
     int getIdMarca ();
     std::string getModelo();
     std::string getDescripcion();
-    double getPrecio ();
+    float getPrecio ();
     int getStock ();
     bool getActivo ();
 
@@ -39,7 +39,7 @@ private:
     int _idMarca;
     char _modelo[15];
     char _descripcion[50];
-    double _precioVenta;
+    float _precioVenta;
     int _stock;
     bool _activo;
 };
